Fix compile warning due to debug message format string.
See #40445

Reported by: Tristan Hoffmann <info@tristanhoffmann.de>

[[originally from svn r15214]]

+++ server/sernet.c
@@ -711,7 +711,8 @@ enum server_events server_sniff_all_input(void)
       if (didget >= 1) {
         buffer[didget-1] = '\0'; /* overwrite newline character */
         didget--;
-        freelog(LOG_DEBUG, "Got line: \"%s\" (%ld, %ld)", buffer, didget, len);
+        freelog(LOG_DEBUG, "Got line: \"%s\" (%ld, %ld)", buffer,
+                (long int) didget, (long int) len);
       }
 #else  /* HAVE_GETLINE */
       buffer = malloc(BUF_SIZE + 1);
